Move some things around to fix web errors
