+++ src/screens/Settings/AppIconSettings/SettingsListItem.web.tsx
@@ -1,0 +1,1 @@
+export function SettingsListItem() {}
