+++ src/screens/Settings/AppearanceSettings.tsx
@@ -1,5 +1,4 @@
-import React, {useCallback, useState} from 'react'
-import {View} from 'react-native'
+import React, {useCallback} from 'react'
 import Animated, {
   FadeInUp,
   FadeOutUp,
@@ -8,25 +7,22 @@ import Animated, {
 } from 'react-native-reanimated'
 import {msg, Trans} from '@lingui/macro'
 import {useLingui} from '@lingui/react'
-import * as DynamicAppIcon from '@mozzius/expo-dynamic-app-icon'
-import {useFocusEffect} from '@react-navigation/native'
 
 import {DISCOVER_DEBUG_DIDS} from '#/lib/constants'
 import {CommonNavigatorParams, NativeStackScreenProps} from '#/lib/routes/types'
 import {isNative} from '#/platform/detection'
 import {useSession} from '#/state/session'
 import {useSetThemePrefs, useThemePrefs} from '#/state/shell'
+import {SettingsListItem as AppIconSettingsListItem} from '#/screens/Settings/AppIconSettings/SettingsListItem'
 import {atoms as a, native, useAlf, useTheme} from '#/alf'
 import * as ToggleButton from '#/components/forms/ToggleButton'
 import {Props as SVGIconProps} from '#/components/icons/common'
 import {Moon_Stroke2_Corner0_Rounded as MoonIcon} from '#/components/icons/Moon'
 import {Phone_Stroke2_Corner0_Rounded as PhoneIcon} from '#/components/icons/Phone'
-import {Shapes_Stroke2_Corner0_Rounded as Shapes} from '#/components/icons/Shapes'
 import {TextSize_Stroke2_Corner0_Rounded as TextSize} from '#/components/icons/TextSize'
 import {TitleCase_Stroke2_Corner0_Rounded as Aa} from '#/components/icons/TitleCase'
 import * as Layout from '#/components/Layout'
 import {Text} from '#/components/Typography'
-import {AppIconImage, useAppIconSets} from './components/AppIcon'
 import * as SettingsList from './components/SettingsList'
 
 type Props = NativeStackScreenProps<CommonNavigatorParams, 'AppearanceSettings'>
@@ -36,23 +32,6 @@ export function AppearanceSettingsScreen({}: Props) {
 
   const {colorMode, darkTheme} = useThemePrefs()
   const {setColorMode, setDarkTheme} = useSetThemePrefs()
-
-  const appIconSets = useAppIconSets()
-  const [currentAppIcon, setCurrentAppIcon] = useState(() =>
-    DynamicAppIcon.getAppIcon(),
-  )
-
-  const icon =
-    appIconSets.defaults.find(i => i.id === currentAppIcon) ??
-    appIconSets.core.find(i => i.id === currentAppIcon) ??
-    appIconSets.defaults[0]
-
-  // refresh current icon when screen is focused
-  useFocusEffect(
-    useCallback(() => {
-      setCurrentAppIcon(DynamicAppIcon.getAppIcon())
-    }, []),
-  )
 
   const onChangeAppearance = useCallback(
     (keys: string[]) => {
@@ -202,19 +181,7 @@ export function AppearanceSettingsScreen({}: Props) {
               {isNative && DISCOVER_DEBUG_DIDS[currentAccount?.did ?? ''] && (
                 <>
                   <SettingsList.Divider />
-
-                  <SettingsList.LinkItem
-                    to="/settings/app-icon"
-                    label={_(msg`App Icon`)}
-                    contentContainerStyle={[a.align_start]}>
-                    <SettingsList.ItemIcon icon={Shapes} />
-                    <View style={[a.flex_1]}>
-                      <SettingsList.ItemText style={[a.pt_xs, a.pb_md]}>
-                        <Trans>App Icon</Trans>
-                      </SettingsList.ItemText>
-                      <AppIconImage icon={icon} size={60} />
-                    </View>
-                  </SettingsList.LinkItem>
+                  <AppIconSettingsListItem />
                 </>
               )}
             </Animated.View>
